validation on candidate page wiht longitued checks etc

+++ system/application/views/projects/closing_documents.php
@@ -35,7 +35,9 @@ $this->load->view('projects/search_form');
 					  <td>{filename}</td>
 					  <td>{attached_on}</td>
 					  <td><a href='{site_url}uploads/{project_id}_{filename}' rel="lyteframe">View</a> | <a href='{site_url}uploads/{project_id}_{filename}'  onClick="window.print();return false">Print</a>
+						<?php if($this->session->userdata('is_admin')) { ?>      
 						 |<a href='{site_url}/index.php/projects/attachment_delete/{id}/{project_id}' class="confirmClick" title="Delete this project" >Delete</a>
+					    <?php } ?>
 					  </td>
 				</tr>	
 				{/closing_attachments}
